fesvr: fix compilation with gcc 13
This patch ports the fix in commit 0a7bb5403d0 to the upstream
riscv-isa-sim repository, whose original commit message is below:

Compiling spike with gcc 13 (for example, included in Fedora 38
prerelease) fails with error:

    In file included from fesvr/syscall.h:6,
                     from fesvr/syscall.cc:4:
    fesvr/device.h:15:30: error: ‘uint64_t’ was not declared in this scope
       15 |   typedef std::function<void(uint64_t)> callback_t;
          |                              ^~~~~~~~

This is due to a gcc header dependency change. See for reference:
https://gcc.gnu.org/gcc-13/porting_to.html#header-dep-changes

This commit explicitly adds the missing <cstdint> header inclusion
to fix this build failure.